Changed unnecessary doubles to ints

+++ RPG/src/Chest.java
@@ -3,7 +3,7 @@ public class Chest extends GameObject{
 
 	private double amountOfGold;
 
-	public Chest(double x, double y, double w, double h, double gold) {
+	public Chest(double x, double y, int w, int h, double gold) {
 		super(x, y, w, h, false, true, 1, "Sprites/chest_empty_open_anim_f0.png");
 		setAmountOfGold(gold);
 	}
