+++ RPG/src/Knight.java
@@ -1,7 +1,7 @@
 
 public class Knight extends Player {
 
-	public Knight(double x, double y, double w, double h) {
+	public Knight(double x, double y, int w, int h) {
 		super(x, y, w, h);
 	}
