+++ RPG/src/Wall.java
@@ -1,7 +1,7 @@
 
 public class Wall extends GameObject {
 
-	public Wall(double x, double y, double w, double h) {
+	public Wall(double x, double y, int w, int h) {
 		super(x, y, w, h, false, true, 1, "/sprites/wall_mid.png"); // uses Environment's constructor #1
 	}
